+++ phonebook-api/src/main/java/com/telran/phonebookapi/controller/ContactController.java
@@ -2,12 +2,10 @@ package com.telran.phonebookapi.controller;
 
 import com.telran.phonebookapi.dto.ContactDto;
 import com.telran.phonebookapi.exception.UserAlreadyExistsException;
-import com.telran.phonebookapi.mapper.AddressMapper;
 import com.telran.phonebookapi.mapper.ContactMapper;
-import com.telran.phonebookapi.mapper.EmailMapper;
-import com.telran.phonebookapi.mapper.PhoneMapper;
 import com.telran.phonebookapi.model.Contact;
 import com.telran.phonebookapi.service.ContactService;
+import com.telran.phonebookapi.service.UserService;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
@@ -16,25 +14,18 @@ import javax.validation.Valid;
 import java.util.List;
 import java.util.stream.Collectors;
 
-
-
 @RestController
 @RequestMapping("/api/contact")
 public class ContactController {
 
     static final String CONTACT_DOES_NOT_BELONG = "Error! This contact doesn't belong this user";
-
+    UserService userService;
     ContactService contactService;
-    AddressMapper addressMapper;
-    PhoneMapper phoneMapper;
-    EmailMapper emailMapper;
     ContactMapper contactMapper;
 
-    public ContactController(ContactService contactService, AddressMapper addressMapper, PhoneMapper phoneMapper, EmailMapper emailMapper, ContactMapper contactMapper) {
+    public ContactController(UserService userService, ContactService contactService, ContactMapper contactMapper) {
+        this.userService = userService;
         this.contactService = contactService;
-        this.addressMapper = addressMapper;
-        this.phoneMapper = phoneMapper;
-        this.emailMapper = emailMapper;
         this.contactMapper = contactMapper;
     }
 
@@ -52,30 +43,37 @@ public class ContactController {
         Contact contact = contactService.getById(id);
         if (!contact.getUser().getEmail().equals(email)) {
             throw new UserAlreadyExistsException(CONTACT_DOES_NOT_BELONG);
-        } else {
-            return ContactDto.builder()
-                    .firstName(contact.getFirstName())
-                    .lastName(contact.getLastName())
-                    .description(contact.getDescription())
-                    .build();
         }
+        return ContactDto.builder()
+                .firstName(contact.getFirstName())
+                .lastName(contact.getLastName())
+                .description(contact.getDescription())
+                .build();
     }
 
-   /* @PutMapping("")
+    @PutMapping("")
     public void editContact(Authentication auth, @Valid @RequestBody ContactDto contactDto) {
         UserDetails userDetails = (UserDetails) auth.getPrincipal();
         String email = userDetails.getUsername();
         Contact contact = contactService.getById(contactDto.id);
         if (!contact.getUser().getEmail().equals(email)) {
             throw new UserAlreadyExistsException(CONTACT_DOES_NOT_BELONG);
-        } else {
+        }
         contactService.edit(contactDto.id, contactDto.firstName, contactDto.lastName, contactDto.description);
     }
 
+
     @DeleteMapping("/{id}")
-    public void removeById(@PathVariable int id) {
+    public void removeById(Authentication auth, @PathVariable int id) {
+        UserDetails userDetails = (UserDetails) auth.getPrincipal();
+        String email = userDetails.getUsername();
+        Contact contact = contactService.getById(id);
+        if (!contact.getUser().getEmail().equals(email)) {
+            throw new UserAlreadyExistsException(CONTACT_DOES_NOT_BELONG);
+        }
         contactService.removeById(id);
-    }*/
+    }
+
 
     @GetMapping("")
     public List<ContactDto> requestAllContactsByUserEmail(Authentication auth) {
@@ -85,19 +83,23 @@ public class ContactController {
                 .collect(Collectors.toList());
     }
 
-   /*@PostMapping("/profile")
-    public void addProfile(@Valid @RequestBody ContactDto contactDto) {
-        contactService.addProfile(contactDto);
+    @GetMapping("/profile")
+    public ContactDto getProfile(Authentication auth) {
+        UserDetails userDetails = (UserDetails) auth.getPrincipal();
+        String email = userDetails.getUsername();
+        Contact myProfile = contactService.getProfile(email);
+        return ContactDto.builder()
+               .firstName(myProfile.getFirstName())
+                .lastName(myProfile.getLastName())
+                .description(myProfile.getDescription())
+                .build();
     }
 
     @PutMapping("/profile")
-    public void editProfile(@Valid @RequestBody ContactDto contactDto) {
-        contactService.editProfile(contactDto);
+    public void editProfile(Authentication auth, @Valid @RequestBody ContactDto myProfileDto) {
+        UserDetails userDetails = (UserDetails) auth.getPrincipal();
+        String email = userDetails.getUsername();
+        contactService.editProfile(email, myProfileDto.firstName, myProfileDto.lastName, myProfileDto.description);
     }
-
-    @PostMapping("/get-profile")
-    public ContactDto getProfile(@Valid @RequestBody UserEmailDto userEmailDto) {
-        return contactService.getProfile(userEmailDto);
-    }*/
 
 }
